requirements.txt
django==3.2.16
djangorestframework==3.14.0
psycopg2-binary==2.9.5
django-debug-toolbar
